Delete old tanStackRoutes.tsx
