--- skribenten-web/frontend/src/pages/Brevvelger/SelectedTemplate.tsx
@@ -1,342 +1,0 @@
-import { css } from "@emotion/react";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { ArrowRightIcon, StarFillIcon, StarIcon } from "@navikt/aksel-icons";
-import { Alert, BodyShort, Button, Heading, Radio, RadioGroup, Select, Tag, VStack } from "@navikt/ds-react";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { useNavigate, useParams, useRouteContext, useSearch } from "@tanstack/react-router";
-import type { AxiosError } from "axios";
-import { useEffect } from "react";
-import { Controller, FormProvider, useForm, useFormContext } from "react-hook-form";
-import { z } from "zod";
-
-import {
-  addFavoritt,
-  deleteFavoritt,
-  getFavoritter,
-  getLetterTemplate,
-  getMineEnheter,
-  orderLetter,
-} from "~/api/skribenten-api-endpoints";
-import { Divider } from "~/components/Divider";
-import { usePreferredLanguage } from "~/hooks/usePreferredLanguage";
-import { redigeringRoute, selectedTemplateRoute } from "~/tanStackRoutes";
-import type { LetterMetadata, OrderLetterRequest } from "~/types/apiTypes";
-import { BrevSystem, SpraakKode } from "~/types/apiTypes";
-import { SPRAAK_ENUM_TO_TEXT } from "~/types/nameMappings";
-
-import { BrevvelgerTabOptions } from "./BrevvelgerPage";
-
-const formValidationSchema = z.object({
-  spraak: z.nativeEnum(SpraakKode, { required_error: "Obligatorisk" }),
-  enhetsId: z.string({ required_error: "Obligatorisk" }).length(4, "Obligatorisk"),
-  isSensitive: z.boolean({ required_error: "Obligatorisk" }),
-});
-
-export function SelectedTemplate() {
-  const { fane } = useSearch({ from: selectedTemplateRoute.id });
-  const { templateId } = useParams({ from: selectedTemplateRoute.id });
-  const { getSakQueryOptions } = useRouteContext({ from: selectedTemplateRoute.id });
-  const sak = useQuery(getSakQueryOptions).data;
-
-  const letterTemplate = useQuery({
-    queryKey: getLetterTemplate.queryKey(sak?.sakType as string),
-    queryFn: () => getLetterTemplate.queryFn(sak?.sakType as string),
-    select: (letterTemplates) =>
-      [...letterTemplates.eblanketter, ...letterTemplates.kategorier.flatMap((kategori) => kategori.templates)].find(
-        (letterMetadata) => letterMetadata.id === templateId,
-      ),
-    enabled: !!sak,
-  }).data;
-
-  if (!letterTemplate) {
-    return <></>;
-  }
-
-  return (
-    <div
-      css={css`
-        display: flex;
-        padding: var(--a-spacing-6) var(--a-spacing-4);
-        flex-direction: column;
-        align-items: flex-start;
-        gap: var(--a-spacing-5);
-        border-right: 1px solid var(--a-gray-400);
-      `}
-    >
-      <FavoriteButton />
-      {fane === BrevvelgerTabOptions.BREVMALER ? (
-        <Brevmal letterTemplate={letterTemplate} />
-      ) : (
-        <Eblankett letterTemplate={letterTemplate} />
-      )}
-    </div>
-  );
-}
-
-function Brevmal({ letterTemplate }: { letterTemplate: LetterMetadata }) {
-  const { templateId, sakId } = useParams({ from: selectedTemplateRoute.id });
-  const { enhetsId } = useSearch({ from: selectedTemplateRoute.id });
-  const navigate = useNavigate({ from: selectedTemplateRoute.id });
-  const { getSakQueryOptions } = useRouteContext({ from: selectedTemplateRoute.id });
-  const sak = useQuery(getSakQueryOptions).data;
-
-  const orderLetterMutation = useMutation<string, AxiosError<Error> | Error, OrderLetterRequest>({
-    mutationFn: orderLetter,
-    onSuccess: (callbackUrl) => {
-      window.open(callbackUrl);
-    },
-  });
-
-  const methods = useForm<z.infer<typeof formValidationSchema>>({
-    defaultValues: {
-      enhetsId,
-      isSensitive: letterTemplate?.brevsystem === BrevSystem.Extream ? undefined : false, // Supply default value to pass validation if Brev is not Doksys
-    },
-    resolver: zodResolver(formValidationSchema),
-  });
-
-  return (
-    <>
-      <LetterTemplateHeading letterTemplate={letterTemplate} />
-      <Heading level="3" size="xsmall">
-        Formål og målgruppe
-      </Heading>
-      <BodyShort size="small">TODO</BodyShort>
-      <Divider />
-      <Heading level="3" size="xsmall">
-        Mottaker (TODO)
-      </Heading>
-      <FormProvider {...methods}>
-        <form
-          css={css`
-            display: flex;
-            flex-direction: column;
-            height: 100%;
-            justify-content: space-between;
-          `}
-          onSubmit={methods.handleSubmit((submittedValues) => {
-            switch (letterTemplate.brevsystem) {
-              case BrevSystem.Extream: {
-                return navigate({ to: redigeringRoute.id, params: { sakId, templateId } });
-              }
-              case BrevSystem.Extream:
-              case BrevSystem.DokSys: {
-                const orderLetterRequest = {
-                  brevkode: letterTemplate.id,
-                  sakId: Number(sakId),
-                  gjelderPid: sak?.foedselsnr ?? "TODO",
-                  ...submittedValues,
-                };
-                return orderLetterMutation.mutate(orderLetterRequest);
-              }
-            }
-          })}
-        >
-          <VStack gap="4">
-            <SelectLanguage letterTemplate={letterTemplate} />
-            <SelectEnhetsId />
-            <SelectSensitivity letterTemplate={letterTemplate} />
-          </VStack>
-
-          <VStack gap="4">
-            {orderLetterMutation.error && <Alert variant="error">{orderLetterMutation.error.message}</Alert>}
-            <Button
-              css={css`
-                width: fit-content;
-              `}
-              icon={<ArrowRightIcon />}
-              iconPosition="right"
-              loading={orderLetterMutation.isPending}
-              size="small"
-              type="submit"
-              variant="primary"
-            >
-              Rediger brev
-            </Button>
-          </VStack>
-        </form>
-      </FormProvider>
-    </>
-  );
-}
-
-function SelectSensitivity({ letterTemplate }: { letterTemplate: LetterMetadata }) {
-  if (letterTemplate.brevsystem !== BrevSystem.Extream) {
-    return <></>;
-  }
-
-  return (
-    <Controller
-      name="isSensitive"
-      render={({ field, fieldState }) => (
-        <RadioGroup
-          legend="Er brevet sensitivt?"
-          {...field}
-          error={fieldState.error?.message}
-          size="small"
-          value={field.value ?? null}
-        >
-          <Radio value>Ja</Radio>
-          <Radio value={false}>Nei</Radio>
-        </RadioGroup>
-      )}
-    />
-  );
-}
-
-function SelectLanguage({ letterTemplate }: { letterTemplate: LetterMetadata }) {
-  const { sakId } = useParams({ from: selectedTemplateRoute.id });
-  const { register, setValue } = useFormContext();
-  const preferredLanguage = usePreferredLanguage(sakId);
-
-  // Update selected language if preferredLanguage was not loaded before form initialization.
-  useEffect(() => {
-    if (preferredLanguage && letterTemplate.spraak.includes(preferredLanguage)) {
-      setValue("spraak", preferredLanguage);
-    }
-  }, [preferredLanguage, setValue, letterTemplate.spraak]);
-
-  return (
-    <Select {...register("spraak")} label="Språk" size="small">
-      {letterTemplate.spraak.map((spraak) => (
-        <option key={spraak} value={spraak}>
-          {SPRAAK_ENUM_TO_TEXT[spraak]} {preferredLanguage === spraak ? "(foretrukket språk)" : ""}
-        </option>
-      ))}
-    </Select>
-  );
-}
-
-function SelectEnhetsId() {
-  const { register, formState } = useFormContext();
-
-  const enheter = useQuery({
-    ...getMineEnheter,
-  }).data;
-
-  return (
-    <Select
-      {...register("enhetsId")}
-      error={formState.errors.enhetsId?.message?.toString()}
-      label="EnhetsId"
-      size="small"
-    >
-      {enheter?.map((enhet) => (
-        <option key={enhet.enhetNr} value={enhet.enhetNr}>{`${enhet.navn} (${enhet.enhetNr})`}</option>
-      ))}
-    </Select>
-  );
-}
-
-function Eblankett({ letterTemplate }: { letterTemplate: LetterMetadata }) {
-  return (
-    <>
-      <LetterTemplateHeading letterTemplate={letterTemplate} />
-      <Heading level="3" size="xsmall">
-        Formål og målgruppe
-      </Heading>
-      <BodyShort size="small">E-blankett</BodyShort>
-      <Divider />
-    </>
-  );
-}
-
-function LetterTemplateHeading({ letterTemplate }: { letterTemplate: LetterMetadata }) {
-  return (
-    <div>
-      <Heading level="2" size="medium">
-        {letterTemplate.name}
-      </Heading>
-      <div
-        css={css`
-          display: flex;
-          align-items: center;
-          gap: var(--a-spacing-3);
-          margin-top: var(--a-spacing-2);
-        `}
-      >
-        <LetterTemplateTags letterTemplate={letterTemplate} />
-      </div>
-    </div>
-  );
-}
-
-function LetterTemplateTags({ letterTemplate }: { letterTemplate: LetterMetadata }) {
-  if (letterTemplate.isEblankett) {
-    return <></>;
-  }
-
-  return (
-    <div>
-      {(() => {
-        switch (letterTemplate.brevsystem) {
-          case BrevSystem.Brevbaker: {
-            return (
-              <Tag size="small" variant="alt2-moderate">
-                Brevbaker
-              </Tag>
-            );
-          }
-          case BrevSystem.Extream: {
-            return (
-              <Tag size="small" variant="alt1-moderate">
-                Extream
-              </Tag>
-            );
-          }
-          case BrevSystem.DokSys: {
-            return (
-              <Tag size="small" variant="alt3-moderate">
-                Doksys
-              </Tag>
-            );
-          }
-        }
-      })()}
-    </div>
-  );
-}
-
-function FavoriteButton() {
-  const { templateId } = useParams({ from: selectedTemplateRoute.id });
-  const queryClient = useQueryClient();
-  const isFavoritt = useQuery({
-    ...getFavoritter,
-    select: (favoritter) => favoritter.includes(templateId),
-  }).data;
-
-  const toggleFavoritesMutation = useMutation<unknown, unknown, string>({
-    mutationFn: (id) => (isFavoritt ? deleteFavoritt(id) : addFavoritt(id)),
-    onSettled: () => queryClient.invalidateQueries({ queryKey: getFavoritter.queryKey }),
-  });
-
-  if (isFavoritt) {
-    return (
-      <Button
-        css={css`
-          width: fit-content;
-        `}
-        icon={<StarFillIcon aria-hidden />}
-        onClick={() => toggleFavoritesMutation.mutate(templateId)}
-        size="small"
-        variant="secondary"
-      >
-        Fjern som favoritt
-      </Button>
-    );
-  }
-
-  return (
-    <Button
-      css={css`
-        width: fit-content;
-      `}
-      icon={<StarIcon aria-hidden />}
-      onClick={() => toggleFavoritesMutation.mutate(templateId)}
-      size="small"
-      variant="secondary"
-    >
-      Legg til som favoritt
-    </Button>
-  );
-}
